+++ client/src/components/MatchHistoryPage/MatchHistoryDisplay/RuneDisplay/DisplayStatMods.ts
@@ -1,4 +1,4 @@
-import { ParticipantDto } from "../../interfaces/MatchV5";
+import { ParticipantDto } from "../../../../types/MatchV5";
 import runes from "../../json/perks.json";
 const DisplayStatMod = (playerData: ParticipantDto, statType: string) => {
   const defense = playerData.perks.statPerks.defense;
